Separating `ActiveAdvancedFilters` as its own component.
